requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy==2.0.21
psycopg2-binary
alembic
python-dotenv
pydantic
